requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
piexif==1.1.3
pillow==10.3.0
Werkzeug==3.0.3
